+++ internal/config/about.go
